+++ packages/eyes-selenium/test/util/version-alias.js
@@ -4,4 +4,7 @@ if (!process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION) {
   process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION = '4'
 }
 
-alias.addAlias('selenium-webdriver', `selenium-webdriver-${process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION}`)
+alias.addAlias(
+  'selenium-webdriver',
+  `selenium-webdriver-${process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION}`,
+)