+++ packages/sdk-coverage-tests/coverage-tests/custom/TestCheckScrollableModal_VG.spec.js
@@ -1,0 +1,25 @@
+'use strict'
+const path = require('path')
+const cwd = process.cwd()
+const spec = require(path.resolve(cwd, 'src/SpecWrappedDriver'))
+const {getEyes, Browsers} = require('../util/TestSetup')
+const {TestCheckScrollableModal} = require('./TestFluentApi_utils')
+
+describe('Coverage tests', () => {
+  let driver, eyes
+
+  afterEach(async () => {
+    await eyes.abortIfNotClosed()
+    await spec.cleanup(driver)
+  })
+
+  beforeEach(async () => {
+    driver = await spec.build({capabilities: Browsers.chrome()})
+    await spec.visit(driver, 'https://applitools.github.io/demo/TestPages/FramesTestPage/')
+    eyes = await getEyes({isVisualGrid: true, branchName: 'v2'})
+    eyes.setMatchTimeout(0)
+  })
+
+  it('TestCheckScrollableModal_VG', () =>
+    TestCheckScrollableModal({testName: 'TestCheckScrollableModal_VG', eyes, driver}))
+})