+++ packages/visual-grid-client/CHANGELOG.md
@@ -3,6 +3,14 @@
 ## Unreleased
 
 
+## 14.4.6 - 2020/6/12
+
+- updated to @applitools/eyes-sdk-core@11.0.7 (from 11.0.6)
+
+## 14.4.5 - 2020/6/11
+
+- updated to @applitools/eyes-sdk-core@11.0.6 (from 11.0.5)
+
  ## 14.4.4 - 2020/6/9
 
 - added chromeEmulationInfo and TS defs for iosDeviceInfo